pyproject.toml
[tool.poetry]
name = "debug_drf_pytest"
version = "0.1.0"
description = ""
authors = ["Etty <beltrami.ester@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.8"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
pytest-django = "^3.9.0"
pytest-coverage = "^0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
